web-app/backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.22.0
python-dotenv>=0.21.0
requests>=2.31.0
qrcode[pil]>=7.4.2
Pillow>=10.3.0
